requirements.txt
django==3.2
django-templated-mail==1.1.1
Pillow==9.1.1
djangorestframework==3.14.0
requests==2.27.1
djoser==2.1.0
django-cors-headers==3.13.0
openpyxl==3.0.10
django-admin-interface==0.24.2
psycopg2-binary==2.9.5
djangorestframework-simplejwt==4.8.0
translate==3.6.1
polib==1.1.1
qrcode==7.4.2
swagger-ui-py==22.7.13
PyYAML==6.0
uritemplate==4.1.1
drf-spectacular==0.26.0
easy-thumbnails==2.8.5
channels==4.0.0
daphne==4.0.0
